DV-161 feat: 필요없는 파일 정리 및 통합, README 파일 작성
