+++ README.md
@@ -7,7 +7,12 @@ Interview King Dev Kim AI Repository
 1. 프로젝트 루트에서 아래 명령어를 실행하여 로컬 서버를 시작합니다. (실행 전에 가상 환경을 활성화하고, 필요한 의존성들이 모두 설치되어 있어야 합니다.)
 
 ```bash
-fastapi dev app/main.py
+poetry shell # 가상환경 활성화
+poetry install # 의존성 설치
+```
+
+```bash
+poetry run fastapi dev app/main.py
 ```
 
 ### Branch 전략
@@ -38,57 +43,29 @@ Git-flow
 - ci: CI/CD 설정 및 스크립트 변경
 - build: 빌드 관련 작업
 
-### 의존성 설치
-
-1.  프로젝트 루트에서 가상 환경을 생성합니다.
-
-```bash
-python -m venv venv
-```
-
-2.  가상 환경을 활성화합니다.
-
-```bash
-source venv/bin/activate
-```
-
-3.  requirements.txt에 있는 라이브러리를 가상 환경에 설치합니다.
-
-```bash
-pip install -r requirements.txt
-```
-
-### 의존성 업데이트
-
-1.  가상 환경을 활성화합니다. (가상 환경 venv이 먼저 생성되어 있어야 합니다)
-
-```bash
-source venv/bin/activate
-```
-
-2.  현재 설치된 라이브러리 목록을 requirements.txt 파일에 업데이트합니다.
-
-```bash
-pip freeze > requirements.txt
-```
-
-### 폴더 구조
-
-- app
-  - ai
-  - logs
-  - models
-
 ### Poetry
 
+의존성 설치
+`poetry install`
+
 패키지 추가
-poetry add python-dotenv
+`poetry add python-dotenv`
 
 dev 패키지 추가
-poetry add pytest -D
-
-가상환경 활성화
-poetry shell
+`poetry add pytest -D`
 
 설치된 패키지 확인
-poetry show
+`poetry show`
+
+가상환경 활성화
+`poetry shell`
+
+가상환경 내에서 명령어 실행
+`poetry run fastapi dev app/main.py # in root directory`
+
+# 폴더 구조 txt 파일 생성
+
+```bash
+brew install tree # macOS
+tree -I 'wandb|**pycache**|\*.log|__pycache__|__init__|folder_structure' > folder_structure.txt
+```
