+++ app/services/interview_service.py
@@ -10,8 +10,9 @@ from app.schemas.evaluation import EvaluationRequest
 from app.schemas.question import QuestionsRequest
 from app.services.evaluation_service import generate_evaluation_prompt
 from app.services.question_service import generate_questions_prompt
-from app.utils.generate import create_file_objects, get_cover_letters_data, get_files_from_s3
+from app.utils.generate import create_file_objects, get_cover_letters_data
 from app.utils.merge import merge_questions_and_answers
+from app.utils.s3 import get_files_from_s3
 
 load_dotenv()
 
@@ -35,10 +36,10 @@ async def generate_interview_questions(user_data: QuestionsRequest):
 
 
 def generate_interview_evaluation(user_data: EvaluationRequest):
+    prompt = generate_evaluation_prompt(user_data)
+
     merged_input = merge_questions_and_answers(user_data.questions, user_data.answers)
     merged_input_str = json.dumps(merged_input, ensure_ascii=False)
-
-    prompt = generate_evaluation_prompt(user_data)
 
     generator = ContentGenerator(user_data=user_data)
     data = generator.invoke(prompt, merged_input_str, "evaluation")
