--- app/utils/logger.py
@@ -1,23 +1,0 @@
-import logging
-
-
-def setup_logger(log_level=logging.INFO, log_file=None):
-    logger = logging.getLogger()
-    logger.setLevel(log_level)
-
-    formatter = logging.Formatter("%(asctime)s - %(name)s - %(levelname)s - %(message)s")
-
-    console_handler = logging.StreamHandler()
-    console_handler.setFormatter(formatter)
-    logger.addHandler(console_handler)
-
-    if log_file:
-        file_handler = logging.FileHandler(log_file)
-        file_handler.setFormatter(formatter)
-        logger.addHandler(file_handler)
-
-    return logger
-
-
-def get_logger():
-    return logging.getLogger()
